+++ Main/Features/ArmyTargeting/ArmyTargetingIoC.cs
@@ -1,0 +1,13 @@
+using DryIoc;
+
+namespace TAOM.Features.ArmyTargeting;
+
+public static class ArmyTargetingIoC
+{
+    public static void RegisterArmyTargetingFeature(IContainer container)
+    {
+        container.Register<IArmyTargetingConfigProvider, ArmyTargetingConfigProvider>(Reuse.Singleton);
+        container.Register<IArmyTargetingSettingsProvider, ArmyTargetingSettingsProvider>(Reuse.Singleton);
+        container.Register<IArmyTargetingService, ArmyTargetingService>(Reuse.Singleton);
+    }
+}